chore: consolidated app stack
