+++ src/frontend/hooks/useNavigationWithTypes.ts
@@ -1,7 +1,7 @@
 import {BottomTabNavigationProp} from '@react-navigation/bottom-tabs';
 import {CompositeNavigationProp, useNavigation} from '@react-navigation/native';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
-import {AppStackList} from '../Navigation/AppStack';
+import {AppStackList} from '../Navigation/Stack';
 import {type HomeTabsList} from '../Navigation/Tab';
 
 export const useNavigationFromRoot = () =>
